feat: Refactored Layout System to Match Figma Design
