+++ public/assets/candle_stick.png
