+++ src/app/component/CourseCard.tsx
