+++ src/app/component/Hero_section.tsx
